+++ pai-dlc-20201203/alibabacloud_pai_dlc20201203/models/_get_job_response_body.py
@@ -14,6 +14,7 @@ class GetJobResponseBody(DaraModel):
         cluster_id: str = None,
         code_source: main_models.GetJobResponseBodyCodeSource = None,
         credential_config: main_models.CredentialConfig = None,
+        custom_envs: List[main_models.GetJobResponseBodyCustomEnvs] = None,
         data_sources: List[main_models.GetJobResponseBodyDataSources] = None,
         display_name: str = None,
         duration: int = None,
@@ -65,6 +66,7 @@ class GetJobResponseBody(DaraModel):
         self.code_source = code_source
         # The access credential configurations.
         self.credential_config = credential_config
+        self.custom_envs = custom_envs
         # The data sources.
         self.data_sources = data_sources
         # The job name.
@@ -161,6 +163,10 @@ class GetJobResponseBody(DaraModel):
             self.code_source.validate()
         if self.credential_config:
             self.credential_config.validate()
+        if self.custom_envs:
+            for v1 in self.custom_envs:
+                 if v1:
+                    v1.validate()
         if self.data_sources:
             for v1 in self.data_sources:
                  if v1:
@@ -208,6 +214,11 @@ class GetJobResponseBody(DaraModel):
 
         if self.credential_config is not None:
             result['CredentialConfig'] = self.credential_config.to_map()
+
+        result['CustomEnvs'] = []
+        if self.custom_envs is not None:
+            for k1 in self.custom_envs:
+                result['CustomEnvs'].append(k1.to_map() if k1 else None)
 
         result['DataSources'] = []
         if self.data_sources is not None:
@@ -355,6 +366,12 @@ class GetJobResponseBody(DaraModel):
         if m.get('CredentialConfig') is not None:
             temp_model = main_models.CredentialConfig()
             self.credential_config = temp_model.from_map(m.get('CredentialConfig'))
+
+        self.custom_envs = []
+        if m.get('CustomEnvs') is not None:
+            for k1 in m.get('CustomEnvs'):
+                temp_model = main_models.GetJobResponseBodyCustomEnvs()
+                self.custom_envs.append(temp_model.from_map(k1))
 
         self.data_sources = []
         if m.get('DataSources') is not None:
@@ -1107,6 +1124,49 @@ class GetJobResponseBodyDataSources(DaraModel):
 
         if m.get('Uri') is not None:
             self.uri = m.get('Uri')
+
+        return self
+
+class GetJobResponseBodyCustomEnvs(DaraModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+        visible: str = None,
+    ):
+        self.key = key
+        self.value = value
+        self.visible = visible
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.key is not None:
+            result['Key'] = self.key
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        if self.visible is not None:
+            result['Visible'] = self.visible
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+
+        if m.get('Visible') is not None:
+            self.visible = m.get('Visible')
 
         return self